#11 - Refactor: extract to its own test file
